worker/requirements.txt
psycopg2 == 2.9.9
python-multipart == 0.0.9
boto3 == 1.35.22
redis == 4.5
pytest == 8.3.3
pillow == 10.4.0
